상태머신 관련 커밋

+++ RoguelightSpeedRun20D/Assets/_/minwoo/01.Scripts/01.Item/03.Class/01.ItemClasses/01.Equip/Weapon.cs
@@ -6,7 +6,7 @@ public class Weapon : Equipment
 {
     int damage;
     bool isRangeAttack;
-    int size;
+    float size;
     float cooltime;
 
     public Weapon(string name, BasicEquipments basicData, WeaponData data)
@@ -22,7 +22,7 @@ public class Weapon : Equipment
 
     public int Damage { get => damage; set => damage = value; }
     public bool IsRangeAttack { get => isRangeAttack; set => isRangeAttack = value; }
-    public int Size { get => size; set => size = value; }
+    public float Size { get => size; set => size = value; }
     public float Cooltime { get => cooltime; set => cooltime = value; }
 
     public override void Equip()
